+++ HarmonyDB.Index/HarmonyDB.Index.Analysis/Models/Loop.cs
@@ -39,7 +39,7 @@ public class Loop
         var bytes = new byte[progression.Length * 3];
         for (var index = 0; index < progression.Span.Length; index++)
         {
-            bytes[index * 3] = (byte)progression.Span[index].RootDelta;
+            bytes[index * 3] = progression.Span[index].RootDelta;
             bytes[index * 3 + 1] = (byte)progression.Span[index].FromType;
             bytes[index * 3 + 2] = (byte)progression.Span[index].ToType;
         }
@@ -58,7 +58,7 @@ public class Loop
 
             int[] CreateIdSequence(int shift) => Shift(shift).Select(x => BitConverter.ToInt32(new byte[]
             {
-                (byte)x.RootDelta,
+                x.RootDelta,
                 (byte)x.FromType,
                 (byte)x.ToType,
                 0,
